requirements.txt
semantic-kernel>=0.9.0
python-dotenv>=1.0.0
mcp>=1.5.0
pydantic>=2.0.0
uvicorn>=0.24.0
typing-extensions>=4.7.0
uv
httpx
